+++ libs/langchain/langchain/output_parsers/format_instructions.py
@@ -46,25 +46,3 @@ Please follow the standard YAML formatting conventions with an indent of 2 space
 ```
 
 Make sure to always enclose the YAML output in triple backticks (```)"""
-
-
-PANDAS_DATAFRAME_FORMAT_INSTRUCTIONS = """The output should be formatted as a string as the operation, followed by a colon, followed by the column or row to be queried on, followed by optional array parameters.
-1. The column names are limited to the possible columns below.
-2. Arrays must either be a comma-separated list of numbers formatted as [1,3,5], or it must be in range of numbers formatted as [0..4].
-3. Remember that arrays are optional and not necessarily required.
-4. If the column is not in the possible columns or the operation is not a valid Pandas DataFrame operation, return why it is invalid as a sentence starting with either "Invalid column" or "Invalid operation".
-
-As an example, for the formats:
-1. String "column:num_legs" is a well-formatted instance which gets the column num_legs, where num_legs is a possible column.
-2. String "row:1" is a well-formatted instance which gets row 1.
-3. String "column:num_legs[1,2]" is a well-formatted instance which gets the column num_legs for rows 1 and 2, where num_legs is a possible column.
-4. String "row:1[num_legs]" is a well-formatted instance which gets row 1, but for just column num_legs, where num_legs is a possible column.
-5. String "mean:num_legs[1..3]" is a well-formatted instance which takes the mean of num_legs from rows 1 to 3, where num_legs is a possible column and mean is a valid Pandas DataFrame operation.
-6. String "do_something:num_legs" is a badly-formatted instance, where do_something is not a valid Pandas DataFrame operation.
-7. String "mean:invalid_col" is a badly-formatted instance, where invalid_col is not a possible column.
-
-Here are the possible columns:
-```
-{columns}
-```
-"""
